requirements.txt
# Multi-Agent Fact-Checking System - Python Dependencies

# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
pydantic>=2.0.0
pyyaml>=6.0

# LLM Providers (Free - Primary)
ollama>=0.3.0

# LLM Providers (Paid - Optional)
openai>=1.0.0
google-generativeai>=0.8.0

# Web Search (Free - Primary)
duckduckgo-search>=6.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
html5lib>=1.1

# Optional: Selenium for JavaScript-heavy sites
selenium>=4.15.0
webdriver-manager>=4.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0

# Logging & Configuration
python-dotenv>=1.0.0
loguru>=0.7.0

# Evaluation & Metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
pre-commit>=3.4.0

# Type Hints
types-requests>=2.31.0
types-pyyaml>=6.0.0
types-beautifulsoup4>=4.12.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Date/Time Utilities
python-dateutil>=2.8.0

# Natural Language Processing
spacy>=3.7.0
nltk>=3.8.0

# Multilingual Support
fasttext>=0.9.2
sentence-transformers>=2.2.0

# Progress Bars
tqdm>=4.66.0

# Retry Logic
tenacity>=8.2.0

# JSON Schema
jsonschema>=4.19.0

# HTTP Rate Limiting
ratelimit>=2.2.0

# Dataset Loading (for benchmarks)
datasets>=2.14.0  # HuggingFace datasets

# Text Processing
regex>=2023.8.8

# Caching
diskcache>=5.6.0

# Command Line Interface
click>=8.1.0
rich>=13.5.0  # for beautiful CLI output

# Environment
python-dotenv>=1.0.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Optional: For PDF generation
# reportlab>=4.0.0
# fpdf2>=2.7.0
